Split out predicates into helper class
